+++ src/lib/event_mapper/impl/__init__.py
@@ -4,5 +4,6 @@ from .glue_workflow_event_mapper import GlueWorkflowEventMapper
 from .glue_data_catalog_event_mapper import GlueDataCatalogEventMapper
 from .glue_crawler_event_mapper import GlueCrawlerEventMapper
 from .step_functions_event_mapper import StepFunctionsEventMapper
+from .lambda_functions_event_mapper import LambdaFunctionsEventMapper
 from .general_aws_event_mapper import EventParsingException
 from .general_aws_event_mapper import ExecutionInfoUrlMixin