+++ src/lib/event_mapper/resource_type_resolver.py
@@ -10,7 +10,12 @@ class ResourceTypeResolver:
             "Crawler": types.GLUE_CRAWLERS,
         },
         "aws.states": {"Execution": types.STEP_FUNCTIONS},
-        "salmon.glue_workflow": {"Glue Workflow State Change": types.GLUE_WORKFLOWS}, # custom processing
+        "salmon.glue_workflow": {
+            "Glue Workflow State Change": types.GLUE_WORKFLOWS
+        },  # custom processing
+        "salmon.lambda": {
+            "Lambda Function Execution State Change": types.LAMBDA_FUNCTIONS
+        },  # custom processing
     }
 
     @staticmethod